refactor: replace useStrategies with react-query
